factor common RPC client code into the RPC client library
